+++ superwall/src/main/java/com/superwall/sdk/web/DeepLinkReferrer.kt
@@ -5,6 +5,7 @@ import android.net.Uri
 import com.android.installreferrer.api.InstallReferrerClient
 import com.android.installreferrer.api.InstallReferrerClient.newBuilder
 import com.android.installreferrer.api.InstallReferrerStateListener
+import com.superwall.sdk.deeplinks.redeemableCode
 import com.superwall.sdk.misc.IOScope
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withTimeout
@@ -112,17 +113,7 @@ class DeepLinkReferrer(
             Result.failure(e)
         }
 
-    override fun handleDeepLink(url: Uri): Result<String> {
-        val failure =
-            Result.failure<String>(UnsupportedOperationException("Link not valid for redemption"))
-        return if (url.host == "superwall" && url.lastPathSegment.equals("redeem")) {
-            url.getQueryParameter("code")?.let {
-                Result.success(it)
-            } ?: failure
-        } else {
-            failure
-        }
-    }
+    override fun handleDeepLink(url: Uri): Result<String> = url.redeemableCode
 
     private fun String.getUrlParams(): Map<String, List<String>> {
         val urlParts = split("\\?".toRegex()).filter(String::isNotEmpty)